Fix handling of Floats, enums and add more tests for 'case' functionality

+++ tests-integration/spark/test/src/SparkTypeTest.scala
@@ -2,7 +2,7 @@ import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.functions.explode
 import org.scalatest.FunSuite
 import sparktests.typetests.SparkJobs
-import org.apache.spark.sql.types.{BooleanType, FloatType, IntegerType, StringType, StructField, StructType}
+import org.apache.spark.sql.types.{BooleanType, DoubleType, IntegerType, StringType, StructField, StructType}
 import org.apache.spark.sql.Row
 
 class SparkTypeTest extends FunSuite {
@@ -44,17 +44,17 @@ class SparkTypeTest extends FunSuite {
   }
 
   test("testMaybeFloat") {
-    val data = Seq(Row(9.99f, "bax"), Row(5.55f, "bar"), Row(null, "baz"))
+    val data = Seq(Row(9.99, "bax"), Row(5.55, "bar"), Row(null, "baz"))
     val schema = StructType(List(
-      StructField("foo", FloatType, true),
+      StructField("foo", DoubleType, true),
       StructField("bar", StringType, false),
     ))
     val rdd = localTestSession.sparkContext.parallelize(data)
     val df = localTestSession.createDataFrame(rdd, schema)
     val result = SparkJobs.testMaybeFloat(df)
     assert(result.count() == 2)
-    assert(result.collect()(0)(0) == 9.99f)
-    assert(result.collect()(1)(0) == 5.55f)
+    assert(result.collect()(0)(0) == 9.99)
+    assert(result.collect()(1)(0) == 5.55)
   }
 
   test("testMaybeInt") {
